[FIX] base: wrong context when user.lang=False
The lang in the context was overriden only when it matched an alias,
e.g. `'en'` => `'en_US'` but it didn't replaced a missing lang by the
default lang.

In addition to fixing the formed problem, we also make sure to fallback
on an installed lang instead of hardcoded 'en_US'.

Also the code that was getting the alias from babel is not needed
anymore as it is present in `request.default_lang()` already.

closes odoo/odoo#110066

X-original-commit: f2523c4acda9ff980f8d78a3d63e6984047c7c01
Signed-off-by: Julien Castiaux <juc@odoo.com>